Establish ability to track which interface was discovered on switch
It may become desirable to use switch topology to disambiguate multiple
NIC mac address collection at some point.  This is some work to prepare
for that.  The downside would be having to make some scenarios that
skip switch lookup have to do switch lookup, just in case.  For such
a scenario, ideally we only do that if we see the possibility of
multiple macs to collect.  For now put it on the backburner as the
scenario of collecting multiple PXE macs and keeping them sorted, but
not using a future built in PXE capability seems less likely.